Added Checkbox and InputField
